refactor: make client almost client again
