+++ Dockerfile
@@ -29,7 +29,7 @@ ENV UV_PYTHON_INSTALL_DIR=/python
 
 # Sync the project without its dev dependencies
 RUN --mount=type=cache,target=/root/.cache/uv \
-    uv sync --locked --no-editable --no-dev --managed-python
+    uv sync --locked --no-editable --no-dev --extra server --managed-python
 
 RUN uv pip install debugpy
