+++ src/blueapi/client/client.py
@@ -15,13 +15,13 @@ from observability_utils.tracing import (
     start_as_current_span,
 )
 
+from blueapi.client.session import SessionCacheManager, SessionManager
 from blueapi.config import (
     ApplicationConfig,
     ConfigLoader,
     MissingStompConfigurationError,
 )
 from blueapi.core.bluesky_types import DataEvent
-from blueapi.service.authentication import SessionCacheManager, SessionManager
 from blueapi.service.model import (
     DeviceModel,
     DeviceResponse,
