+++ src/blueapi/service/__init__.py
@@ -1,4 +1,3 @@
-from .authentication import SessionManager
 from .model import DeviceModel, PlanModel
 
-__all__ = ["PlanModel", "DeviceModel", "SessionManager"]
+__all__ = ["PlanModel", "DeviceModel"]
